updating to cpp14 version

+++ src/USER-DEMSI/atom_vec_demsi.cpp
@@ -1393,9 +1393,9 @@ int AtomVecDemsi::write_vel_hybrid(FILE *fp, double *buf)
    return # of bytes of allocated memory
 ------------------------------------------------------------------------- */
 
-bigint AtomVecDemsi::memory_usage()
+double AtomVecDemsi::memory_usage()
 {
-  bigint bytes = 0;
+  double bytes = 0;
 
   if (atom->memcheck("tag")) bytes += memory->usage(tag,nmax);
   if (atom->memcheck("type")) bytes += memory->usage(type,nmax);
